ported r81136 [substr<-(., beg, <neg.>)] from trunk
git-svn-id: https://svn.r-project.org/R/branches/R-4-1-branch@81241 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/character.c
@@ -674,7 +674,7 @@ SEXP attribute_hidden do_substrgets(SEXP call, SEXP op, SEXP args, SEXP env)
 	    ss = CHAR(el);
 	    slen = strlen(ss);
 	    if (start < 1) start = 1;
-	    if (stop > slen) stop = (int) slen; /* SBCS optimization */
+	    if (stop > (int) slen) stop = (int) slen; /* SBCS optimization */
 	    if (start > stop) {
 		/* just copy element across */
 		SET_STRING_ELT(s, i, STRING_ELT(x, i));
